Move stateless functions into utility modules
