+++ lib/DataHarmonizer.js
@@ -1,8 +1,18 @@
 import { isMatch } from 'date-fns';
-import { saveAs } from 'file-saver';
 import Handsontable from 'handsontable';
 import $ from 'jquery';
-import * as XLSX from 'xlsx/xlsx.mjs';
+import { utils as XlsxUtils, read as xlsxRead } from 'xlsx/xlsx.mjs';
+
+import { wait, isValidHeaderRow } from './utils/general';
+import { readFileAsync, updateSheetRange } from './utils/files';
+import { changeCase, fieldUnitBinTest } from './utils/fields';
+import {
+  checkProvenance,
+  itemCompare,
+  testNumericRange,
+  validateValAgainstVocab,
+  validateValsAgainstVocab,
+} from './utils/validation';
 
 import 'handsontable/dist/handsontable.full.css';
 import './data-harmonizer.css';
@@ -335,10 +345,10 @@ class DataHarmonizer {
    *
    * @param {String} file_name: File in templates/[current schema]/ to load.
    */
-  async openFile(dh, file_name) {
+  async openFile(file_name) {
     try {
-      let contentBuffer = await dh.readFileAsync(file_name);
-      dh.loadSpreadsheetData(contentBuffer);
+      let contentBuffer = await readFileAsync(file_name);
+      this.loadSpreadsheetData(contentBuffer);
     } catch (err) {
       console.log(err);
     }
@@ -721,7 +731,7 @@ class DataHarmonizer {
    */
   async runBehindLoadingScreen(fn, args = []) {
     await $('#loading-screen').show('fast', 'swing');
-    await this.wait(200); // Enough of a visual cue that something is happening
+    await wait(200); // Enough of a visual cue that something is happening
     if (args.length) {
       await fn.apply(this, args);
     } else {
@@ -731,41 +741,19 @@ class DataHarmonizer {
     return;
   }
 
-  // wait ms milliseconds
-  wait(ms) {
-    return new Promise((r) => setTimeout(r, ms));
-  }
   /***************************** PRIVATE functions *************************/
 
-  // https://simon-schraeder.de/posts/filereader-async/
-  readFileAsync(file) {
-    return new Promise((resolve, reject) => {
-      let reader = new FileReader();
-
-      reader.onload = () => {
-        resolve(reader.result);
-      };
-
-      reader.onerror = reject;
-
-      reader.readAsBinaryString(file);
-      //reader.readAsArrayBuffer(file);
-    });
-  }
-
   loadSpreadsheetData(data) {
-    const workbook = XLSX.read(data, {
+    const workbook = xlsxRead(data, {
       type: 'binary',
       raw: true,
       cellDates: true, // Ensures date formatted as  YYYY-MM-DD dates
       dateNF: 'yyyy-mm-dd', //'yyyy/mm/dd;@'
     });
 
-    const worksheet = this.updateSheetRange(
-      workbook.Sheets[workbook.SheetNames[0]]
-    );
+    const worksheet = updateSheetRange(workbook.Sheets[workbook.SheetNames[0]]);
 
-    const matrix = XLSX.utils.sheet_to_json(worksheet, {
+    const matrix = XlsxUtils.sheet_to_json(worksheet, {
       header: 1,
       raw: false,
       range: 0,
@@ -778,30 +766,6 @@ class DataHarmonizer {
     } else {
       this.launchSpecifyHeadersModal(matrix);
     }
-  }
-
-  /**
-   * Improve `XLSX.utils.sheet_to_json` performance for Libreoffice Calc files.
-   * Ensures sheet range is accurate. See
-   * https://github.com/SheetJS/sheetjs/issues/764 for more detail.
-   * @param {Object} worksheet SheetJs object.
-   * @returns {Object} SheetJs worksheet with correct range.
-   */
-  updateSheetRange(worksheet) {
-    const range = { s: { r: 20000000, c: 20000000 }, e: { r: 0, c: 0 } };
-    Object.keys(worksheet)
-      .filter((x) => {
-        return x.charAt(0) !== '!';
-      })
-      .map(XLSX.utils.decode_cell)
-      .forEach((x) => {
-        range.s.c = Math.min(range.s.c, x.c);
-        range.s.r = Math.min(range.s.r, x.r);
-        range.e.c = Math.max(range.e.c, x.c);
-        range.e.r = Math.max(range.e.r, x.r);
-      });
-    worksheet['!ref'] = XLSX.utils.encode_range(range);
-    return worksheet;
   }
 
   /**
@@ -837,7 +801,7 @@ class DataHarmonizer {
       $('#specify-headers-modal').modal('show');
       $('#specify-headers-confirm-btn').click(() => {
         const specifiedHeaderRow = parseInt($('#specify-headers-input').val());
-        if (!this.isValidHeaderRow(matrix, specifiedHeaderRow)) {
+        if (!isValidHeaderRow(matrix, specifiedHeaderRow)) {
           $('#specify-headers-err-msg').show();
         } else {
           const mappedMatrixObj = self.mapMatrixToGrid(
@@ -881,15 +845,6 @@ class DataHarmonizer {
       return 2;
     }
     return false;
-  }
-
-  /**
-   * Validates `$('#specify-headers-input')` input.
-   * @param {Array<Array<String>>} matrix
-   * @param {number} row 1-based index used to indicate header row in matrix.
-   */
-  isValidHeaderRow(matrix, row) {
-    return Number.isInteger(row) && row > 0 && row <= matrix.length;
   }
 
   /**
@@ -971,71 +926,6 @@ class DataHarmonizer {
       matrix: [...expectedHeaders, ...mappedDataRows],
       unmappedHeaders: unmappedHeaders,
     };
-  }
-
-  /**
-   * Download matrix to file.
-   * Note that BOM and UTF-8 can create problems on some systems when importing
-   * file.  See "Supported Output Formats" and "UTF-16 Unicode Text" sections of
-   * https://reactian.com/sheetjs-community-edition-spreadsheet-data-toolkit/
-   * and https://github.com/SheetJS/sheetjs
-   * Solution at bottom of: https://github.com/SheetJS/sheetjs/issues/943
-   * The "Comma Separated Values" format is actually UTF-8 with BOM prefix.
-   * @param {Array<Array<String>>} matrix Matrix to download.
-   * @param {String} baseName Basename of downloaded file.
-   * @param {String} ext Extension of downloaded file.
-   */
-  exportFile(matrix, baseName, ext) {
-    const worksheet = XLSX.utils.aoa_to_sheet(matrix);
-    const workbook = XLSX.utils.book_new();
-    let csv;
-    let blob;
-    XLSX.utils.book_append_sheet(workbook, worksheet, 'Sheet1');
-    switch (ext) {
-      case 'xlsx':
-        XLSX.writeFile(workbook, `${baseName}.xlsx`);
-        break;
-      case 'xls':
-        XLSX.writeFile(workbook, `${baseName}.xls`);
-        break;
-      case 'tsv':
-        XLSX.writeFile(workbook, `${baseName}.tsv`, {
-          bookType: 'csv',
-          FS: '\t',
-        });
-        break;
-      case 'csv':
-        XLSX.writeFile(workbook, `${baseName}.csv`, {
-          bookType: 'csv',
-          FS: ',',
-        });
-        break;
-      case 'tsv (UTF-16)':
-        XLSX.writeFile(workbook, `${baseName}.tsv`, {
-          bookType: 'txt',
-          FS: '\t',
-        });
-        break;
-      case 'csv (UTF-16)':
-        XLSX.writeFile(workbook, `${baseName}.csv`, {
-          bookType: 'txt',
-          FS: ',',
-        });
-        break;
-      case 'csv (UTF-8, no BOM)':
-        //Customization: skips BOM prefix '\uFEFF'
-        csv = XLSX.utils.sheet_to_csv(worksheet, { FS: ',' });
-        blob = new Blob([csv], { type: 'text/plain;charset=UTF-8' });
-        //A FileSaver module call, avoids {autoBom: true} parameter
-        saveAs(blob, `${baseName}.csv`);
-        break;
-      case 'csv (ASCII)':
-        //Customization: skips BOM prefix, as above.
-        csv = XLSX.utils.sheet_to_csv(worksheet, { FS: ',' });
-        blob = new Blob([csv], { type: 'text/plain;charset=us-ascii' });
-        saveAs(blob, `${baseName}.csv`);
-        break;
-    }
   }
 
   /**
@@ -1737,7 +1627,7 @@ class DataHarmonizer {
 
     // Test field against capitalization change.
     if (field.capitalize && change[3] && change[3].length > 0)
-      change[3] = this.changeCase(change[3], field);
+      change[3] = changeCase(change[3], field);
 
     // Rules that require a particular column following and/or preceeding
     // current one.
@@ -1823,31 +1713,6 @@ class DataHarmonizer {
   }
 
   /**
-   * Modify a string to match specified case.
-   * @param {String} val String to modify.
-   * @param {String} capitalize Case to modify string to; one of `'UPPER'`,
-   *     `'lower'` or `'Title'`.
-   * @return {String} String with modified case.
-   */
-  changeCase(val, field) {
-    switch (field.structured_pattern.syntax) {
-      case '{lower_case}':
-        val = val.toLowerCase();
-        break;
-      case '{UPPER_CASE}':
-        val = val.toUpperCase();
-        break;
-      case '{Title_Case}':
-        val = val
-          .split(' ')
-          .map((w) => w[0].toUpperCase() + w.substr(1).toLowerCase())
-          .join(' ');
-        break;
-    }
-    return val;
-  }
-
-  /**
    * Modify matrix data for grid according to specified rules.
    * This is useful when calling `hot.loadData`, as cell changes from said method
    * are not recognized by `afterChange`.
@@ -1865,7 +1730,7 @@ class DataHarmonizer {
       if (field.capitalize) {
         for (let row = 0; row < matrix.length; row++) {
           if (!matrix[row][col]) continue;
-          matrix[row][col] = this.changeCase(matrix[row][col], field);
+          matrix[row][col] = changeCase(matrix[row][col], field);
         }
       }
 
@@ -1894,7 +1759,7 @@ class DataHarmonizer {
                 );
               }
             }
-          } else if (this.fieldUnitBinTest(fields, col)) {
+          } else if (fieldUnitBinTest(fields, col)) {
             // 2 specifies bin offset
             this.binChangeTest(matrix, 0, col, fields, 2, triggered_changes);
           }
@@ -1939,19 +1804,6 @@ class DataHarmonizer {
     }
     // Update changed value (note "change" object overrides triggered_changes)
     return dateParts.join('-');
-  }
-
-  /**
-   * Test to see if col's field is followed by [field unit],[field bin] fields
-   * @param {Object} fields See `data.js`.
-   * @param {Integer} column of numeric field.
-   */
-  fieldUnitBinTest(fields, col) {
-    return (
-      fields.length > col + 2 &&
-      fields[col + 1].title == fields[col].title + ' unit' &&
-      fields[col + 2].title == fields[col].title + ' bin'
-    );
   }
 
   /**
@@ -2076,7 +1928,7 @@ class DataHarmonizer {
         // 1st row of provenance datatype field is forced to have a
         // 'DataHarmonizer Version: 0.13.0' etc. value.  Change happens silently.
         if (datatype === 'Provenance') {
-          this.checkProvenance(provenanceChanges, cellVal, row, col);
+          checkProvenance(provenanceChanges, VERSION_TEXT, cellVal, row, col);
         }
 
         let valid = false;
@@ -2097,34 +1949,33 @@ class DataHarmonizer {
               parsedInt = parseInt(cellVal, 10);
               valid = !isNaN(cellVal);
               valid &= parsedInt.toString() === cellVal;
-              valid &= this.testNumericRange(parsedInt, field);
+              valid &= testNumericRange(parsedInt, field);
               break;
 
             case 'xsd:nonNegativeInteger':
               parsedInt = parseInt(cellVal, 10);
               valid = !isNaN(cellVal) && parsedInt >= 0;
               valid &= parsedInt.toString() === cellVal;
-              valid &= this.testNumericRange(parsedInt, field);
+              valid &= testNumericRange(parsedInt, field);
               break;
 
             case 'xsd:float':
               parsedFloat = parseFloat(cellVal);
               valid = !isNaN(cellVal) && parsedFloat == cellVal;
-              valid &= this.testNumericRange(parsedFloat, field);
+              valid &= testNumericRange(parsedFloat, field);
               break;
 
             case 'xsd:double':
               // NEED DOUBLE RANGE VALIDATION
               parsedFloat = parseFloat(cellVal);
               //valid = !isNaN(cellVal) && regexDouble.test(cellVal);
-              valid &=
-                !isNaN(cellVal) && this.testNumericRange(parsedFloat, field);
+              valid &= !isNaN(cellVal) && testNumericRange(parsedFloat, field);
               break;
 
             case 'xsd:decimal':
               parsedDec = parseFloat(cellVal);
               valid = !isNaN(cellVal) && regexDecimal.test(cellVal);
-              valid &= this.testNumericRange(parsedDec, field);
+              valid &= testNumericRange(parsedDec, field);
               break;
 
             // XML Boolean lexical space accepts true, false, and also 1
@@ -2190,7 +2041,7 @@ class DataHarmonizer {
             field.flatVocabulary
           ) {
             if (field.multivalued === true) {
-              const [vocabValid, update] = this.validateValsAgainstVocab(
+              const [vocabValid, update] = validateValsAgainstVocab(
                 cellVal,
                 field
               );
@@ -2198,7 +2049,7 @@ class DataHarmonizer {
               if (update)
                 this.hot.setDataAtCell(row, col, update, 'thisChange');
             } else {
-              const [vocabValid, update] = this.validateValAgainstVocab(
+              const [vocabValid, update] = validateValAgainstVocab(
                 cellVal,
                 field
               );
@@ -2254,53 +2105,6 @@ class DataHarmonizer {
   }
 
   /**
-   * Test cellVal against "DataHarmonizer provenance: vX.Y.Z" pattern and if it
-   * needs an update, do so.
-   * @param {Array} provenanceChanges array of provenance updates
-   * @param {Object} cellVal field value to be tested.
-   * @param {Integer} row index of data
-   * @param {Integer} column index of data
-   */
-  checkProvenance(provenanceChanges, cellVal, row, col) {
-    if (!cellVal) {
-      provenanceChanges.push([row, col, VERSION_TEXT]);
-      return;
-    }
-    // Most of the time this is the first return point.
-    if (cellVal === VERSION_TEXT) return;
-
-    if (cellVal.substring(0, 14) !== 'DataHarmonizer') {
-      provenanceChanges.push([row, col, VERSION_TEXT + ';' + cellVal]);
-      return;
-    }
-    // At this point we have a leading "DataHarmonizer v..." string
-    let splitVal = cellVal.split(';', 2);
-
-    if (splitVal.length == 1) provenanceChanges.push([row, col, VERSION_TEXT]);
-    else provenanceChanges.push([row, col, VERSION_TEXT + ';' + splitVal[1]]);
-
-    return;
-  }
-
-  /**
-   * Test a given number against an upper or lower range, if any.
-   * @param {Number} number to be compared.
-   * @param {Object} field that contains min and max limits.
-   * @return {Boolean} validity of field.
-   */
-  testNumericRange(number, field) {
-    if (field.minimum_value !== '') {
-      if (number < field.minimum_value) {
-        return false;
-      }
-    }
-    if (field.maximum_value !== '') {
-      if (number > field.maximum_value) return false;
-    }
-    return true;
-  }
-
-  /**
    * Test a given date against an upper or lower range, if any.
    * @param {Date} date to be compared.
    * @param {Object} field that contains min and max limits.
@@ -2321,18 +2125,18 @@ class DataHarmonizer {
             // Signals lookup expressions:
             if (c_item[0] === '{') {
               if (c_item === '{today}') {
-                if (self.itemCompare(jsDate, TODAY, ptr)) return false;
+                if (itemCompare(jsDate, TODAY, ptr)) return false;
               } else {
                 let field = c_item.substr(1, c_item.length - 2);
                 let col = columnIndex[field];
                 let lookup_item = self.hot.getDataAtCell(row, col);
                 if (lookup_item !== '')
-                  if (self.itemCompare(jsDate, new Date(lookup_item), ptr))
+                  if (itemCompare(jsDate, new Date(lookup_item), ptr))
                     return false;
               }
             } else {
               // Assumes this is just a constant date string.
-              if (self.itemCompare(jsDate, new Date(c_item), ptr)) return false;
+              if (itemCompare(jsDate, new Date(c_item), ptr)) return false;
             }
           }
         }
@@ -2340,72 +2144,6 @@ class DataHarmonizer {
     }
 
     return true;
-  }
-
-  /**
-   * Simplifies logic to compare number or date ranges where test limit
-   * is either min_inclusive or max_inclusive
-   * @param {Date or Number} item_1 First value to compare
-   * @param {Date or Number} item_2 Second value to compare
-   * @param {Boolean} gt Type of comparison: 0 = > , 1 = <
-   * @return {Boolean} Result of comparison
-   */
-  itemCompare(item_1, item_2, gt) {
-    if (gt == 1) return item_1 > item_2;
-    return item_1 < item_2;
-  }
-
-  /**
-   * Validate a value against an array of source values.
-   * FUTURE: optimize - to precompile lowercased sources.
-   * @param {String} val Cell value.
-   * @param {Object} field Field to look for flatVocabulary value in.
-   * @return {Array<Boolean><Boolean/String>}
-   *         [false, false] `delimited_string` does not match `source`,
-   *         [true, false] `delimited_string` matches `source` exactly,
-   *         [true, string] `delimited_string` matches`source` but formatting needs change
-   */
-  validateValAgainstVocab(value, field) {
-    let valid = false;
-    let update = false;
-    if (value) {
-      const trimmedVal = value.trim().toLowerCase();
-      const ptr = field.flatVocabularyLCase.indexOf(trimmedVal);
-      if (ptr >= 0) {
-        valid = true;
-        // Normalised value being suggested for update
-        if (value != field.flatVocabulary[ptr])
-          update = field.flatVocabulary[ptr];
-      }
-    }
-    return [valid, update];
-  }
-
-  /**
-   * Validate csv values against an array of source values.  Leading/tailing
-   * whitespace and case are ignored in validation, but returned value will be
-   * a suggested update to one or more values if any differ in capitalization.
-   * @param {String} delimited_string of values to validate.
-   * @param {Object} field to validate values against.
-   * @return {Array<Boolean><Boolean/String>}
-   *         [false, false] If some value in `delimited_string` is not in `source`,
-   *         [true, false] If every value in `delimited_string` is exactly in `source`,
-   *         [true, string] If every value in `delimited_string` is in `source` but formatting needs change
-   */
-  validateValsAgainstVocab(delimited_string, field) {
-    const self = this;
-    let update_flag = false;
-    let value_array = delimited_string.split(';');
-    value_array.forEach(function (value, index) {
-      const [valid, update] = self.validateValAgainstVocab(value, field);
-      if (!valid) return [false, false];
-      if (update) {
-        update_flag = true;
-        value_array[index] = update;
-      }
-    });
-    if (update_flag) return [true, value_array.join(';')];
-    else return [true, false];
   }
 }
