+++ lib/Toolbar.js
@@ -3,7 +3,9 @@ import 'bootstrap/js/dist/carousel';
 import 'bootstrap/js/dist/dropdown';
 import 'bootstrap/js/dist/modal';
 import '@selectize/selectize';
-import '@selectize/selectize/dist/css/selectize.bootstrap4.css'
+import '@selectize/selectize/dist/css/selectize.bootstrap4.css';
+
+import { exportFile } from './utils/files';
 
 import template from './toolbar.html';
 import gettingStarted from './toolbarGettingStarted';
@@ -39,7 +41,7 @@ class Toolbar {
     // will be used.
     this.updateTemplateOptions();
     if (options.templatePath) {
-      const [folder, template] = options.templatePath.split('/')
+      const [folder, template] = options.templatePath.split('/');
       if (folder in this.menu && template in this.menu[folder]) {
         this.$selectTemplate.val(options.templatePath);
       }
@@ -76,7 +78,7 @@ class Toolbar {
         $('#open-error-modal').modal('show');
       } else {
         dh.invalid_cells = {};
-        dh.runBehindLoadingScreen(dh.openFile, [dh, file]);
+        dh.runBehindLoadingScreen(dh.openFile, [file]);
         $('#file_name_display').text(file.name);
       }
       // Allow consecutive uploads of the same file
@@ -106,11 +108,8 @@ class Toolbar {
       try {
         const baseName = $('#base-name-save-as-input').val();
         const ext = $('#file-ext-save-as-select').val();
-        const matrix = [
-          ...dh.getFlatHeaders(),
-          ...dh.getTrimmedData(this.dh.hot),
-        ];
-        dh.runBehindLoadingScreen(dh.exportFile, [matrix, baseName, ext]);
+        const matrix = [...dh.getFlatHeaders(), ...dh.getTrimmedData()];
+        dh.runBehindLoadingScreen(exportFile, [matrix, baseName, ext]);
         $('#save-as-modal').modal('hide');
       } catch (err) {
         $('#save-as-err-msg').text(err.message);
@@ -133,7 +132,7 @@ class Toolbar {
       if (exportFormat in dh.export_formats) {
         const format = dh.export_formats[exportFormat];
         let outputMatrix = format.method(dh);
-        dh.runBehindLoadingScreen(dh.exportFile, [
+        dh.runBehindLoadingScreen(exportFile, [
           outputMatrix,
           baseName,
           format.fileType,
@@ -302,7 +301,10 @@ class Toolbar {
     }
 
     // Settings -> Jump to...
-    const options = Object.keys(columnCoordinates).map(col => ({text: col, value: col}));
+    const options = Object.keys(columnCoordinates).map((col) => ({
+      text: col,
+      value: col,
+    }));
     const $jumpToInput = $('#jump-to-input');
     if ($jumpToInput[0].selectize) {
       $jumpToInput[0].selectize.destroy();
@@ -316,7 +318,7 @@ class Toolbar {
       this.dh.scrollTo(0, columnX);
       $('#jump-to-modal').modal('hide');
       $jumpToInput[0].selectize.clear();
-    })
+    });
     $('#jump-to-modal').on('shown.bs.modal', () => {
       $jumpToInput[0].selectize.open();
     });
@@ -340,7 +342,7 @@ class Toolbar {
     $('#fill-button').on('click', () => {
       this.dh.fillColumn($fillColumnInput.val(), $fillValueInput.val());
       $fillColumnInput[0].selectize.clear();
-      $fillValueInput.val('')
+      $fillValueInput.val('');
     });
   }
 }
