+++ lib/toolbarGettingStarted.js
@@ -1,7 +1,7 @@
 // TODO: these images add quite a bit of weight to the final bundles. Not
-// necessarily a problem for the web bundle because it is designed to be 
+// necessarily a problem for the web bundle because it is designed to be
 // loaded locally, but it may cause a bit of concern for users of the
-// library bundle. I wonder if this type of documentation should be 
+// library bundle. I wonder if this type of documentation should be
 // externalized anyway to make it easier to update and customize.
 import editCopyPasteDelete from './images/editCopyPasteDelete.gif';
 import toggleRequiredCols from './images/toggleRequiredCols.gif';
